Add archive names for common modules
This causes a crazy bug when packaging for distribution (such as the reference validator and legacy-printer) where two jars of the same name would overwrite each other. For the legacy printer, the JDK one would win over the common one. But, for whatever reason, for the validator the common one would win over the JDK one causing NCDFEs.